fix(scalars): move the utility functions to the correct file
